requirements.txt
Requirements:
Flask>=3.0
twilio>=8.0.0
llama-index
pandas>=2.3.3
openai>=0.27.0
python-dotenv
mcp>=1.18.0
pytest-asyncio
google-generativeai>=0.3.0
llama-index-llms-gemini